Add support for DW_AT_main_subprogram
I'd like the Rust compiler to emit DW_AT_main_subprogram in its DWARF
output, so that debuggers can find the user-provided main entry point.
See https://github.com/rust-lang/rust/issues/32620

This patch adds support for DW_AT_main_subprogram to LLVM to support
this effort.

A variant of this patch was accepted by upstream LLVM; there,
DebugInfoFlags.def already has a flag with value 1<<20, which is why
this patch uses 1<<21 -- this will avoid the need to change the Rust
compiler when a new LLVM is imported.