use -O3 with some optimizations disabled on sparc w/ gcc
A previous commit changed it so that rust's private copy of llvm was
always built with -O0 on sparc, as initially that was the only option
that produced a working compiler.  After further investigation, it was
discovered that applying -fno-delayed-branch and -fno-omit-frame-pointer
was sufficient to avoid all of the code generation issues initially
encountered while retaining -O3.  Investigation with gcc upstream
continues.