try to improve the pass manager order
This changes the order of the GVNs a bit.

I removed @eddyb's GVN added in #35662 - I checked the original
examples, and it does not seem to do anything. Instead of it, I added a
GVN to the *start* of the optimization pipeline, to remove junk.

I moved the second GVN to after memcpyopt, so that we'll have a GVN
after MemCpy optimizations - we already have a GVN "before" it.

Also, I duplicated IndVarSimplify instead of adding a SimplifyCfg -
looking at benchmarks, it should have the same performance impact.