Add support for writing HermitCore (https://hermitcore.org) ELF binaries.
HermitCore is a POSIX-compatible kernel for running a single application in
an isolated environment to get maximum performance and predictable runtime
behavior. It can either be used bare-metal on hardware or a VM (Unikernel)
or side by side to an existing Linux system (Multikernel).
Due to the latter feature, HermitCore binaries are marked with a different
OS ABI to let the Linux ELF loader distinguish them from regular Unix/Linux
binaries and load them using the HermitCore "proxy" tool.